Removed Total Functions
I went in and removed the functions that would have to be put in for
each individual item. This wouldn't work due to how you would have
to write new functions when adding items. Now I'm going to go back
in and try to recreate the original functions so that it works with JS
data.