Updated My Form
I went into the HTML that would become the form once the user had
checked out and add in a lot of failsafes and other things to make
the expericence both easier for me and the user. This included a
selector for the states,  patterns and other things which just
protected from obscure or wrong inputs.