Fixed Rent Button Issue
I went in and fixed an issue where an user could rent a vehicle
without actually clicking the rent button, but by clicking on the div
that the vehicle was in. This was due to my forEach being targeted to
the wrong thing. I also went in and added a few more cars to
my inventory for later use.