Made A Sticky Check Out Button
I went in and made my checkout button stay at the bottom of the
page no matter what. This is so that the user doesn't have to fully
scroll to the bottom to check out, or to see where to check out.
I also made a few minor adjustments such as the cart and total at
the top of the navbar and other minor fixes.