Made Sticky Header
I went in and made my header stick to the top of the screen at
all times, so the user doesn't have to scroll up to see the total
or their cart. Also with this, I made more adjustment to the
way my page formats to screens, especially for phones.