Added Vehicles
I added in some more cars, so it fill in the site better. Another
reason for this was the addition of the sticky header and  checkout
button. Now that the user doesn't have to scroll to the top and
bottom to see their total, or checkout, I can add more without
them having issues.