Simplfied JS Functions and Fixed Out of Stock Function
I went in and broke apart a few of my functions, so that they
could be more easily updated and read. I also fixed the rent
button, so that when it hits zero the button becomes out of
stock. This is better than before when you'd have to click the
button again to get it to do so.